Add files via upload
Simple starting pipeline. Here, each node tries to separately predict if the peptide is binding or not, the final decision is based on the sum of the predictions. Special here is also that weight_decay was needed to keep the GNN stable. For now I only did random selection of the validation set, not the clusters (although I added the cluster-ID under 'score' to the hdf5 file so it is possible, please download the newest version of the dataset from the snellius server).
To improve the model I think we should look into input-feature representation/normalization. Besides the NAN and inf values we could change the representations (log mean/std manipulation, for example). The features are very inconsistent, some have one-hot encoding while others are still using integers. After that we should check which features make the model better/worse. 
These hdf5 are now also only using residue level nodes, which means 9 nodes per peptide. We should also built the atom based hdf5 files.